refactor: pull Auth0AppRoot out of index.tsx
