+++ apps/web-game/src/auth/auth0.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+
+export function Auth0AppRoot({ children }: { children: React.ReactNode }) {
+	const onRedirectCallback = React.useCallback((appState: any) => {
+		window.location.href = appState?.returnTo || window.location.pathname;
+	}, []);
+
+	const Auth0AuthProvider = React.useMemo(
+		// eslint-disable-next-line @typescript-eslint/no-var-requires
+		() => require("@cc-web/auth/auth0/provider").Auth0AuthProvider,
+		[]
+	);
+
+	return (
+		<Auth0AuthProvider onRedirectCallback={onRedirectCallback}>
+			{children}
+		</Auth0AuthProvider>
+	);
+}
