+++ apps/web-game/src/index.tsx
@@ -8,26 +8,9 @@ import { AUTH0_ENABLED } from "@cc-web/auth/auth0/config";
 import { GuestAuthProvider } from "@cc-web/auth/guest/provider";
 
 import { App } from "./app";
+import { Auth0AppRoot } from "./auth/auth0";
 import { useGuestMode } from "./guest";
 import { createAppStore } from "./store";
-
-const Auth0AppRoot = ({ children }: { children: React.ReactNode }) => {
-	const onRedirectCallback = React.useCallback((appState: any) => {
-		window.location.href = appState?.returnTo || window.location.pathname;
-	}, []);
-
-	const Auth0AuthProvider = React.useMemo(
-		// eslint-disable-next-line @typescript-eslint/no-var-requires
-		() => require("@cc-web/auth/auth0/provider").Auth0AuthProvider,
-		[]
-	);
-
-	return (
-		<Auth0AuthProvider onRedirectCallback={onRedirectCallback}>
-			{children}
-		</Auth0AuthProvider>
-	);
-};
 
 const AppRoot = () => {
 	const store = createAppStore();
